pyproject.toml
[tool.poetry]
name = "challengeapi"
version = "0.1"
description = ""
authors = ["Dmytro Klymenko <119627462+leegosx@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
faker = "^19.13.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
passlib = "^1.7.4"
pytest = "^8.2.2"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
